test: consolidate worker/api db mock into shared test-fixtures
The mocked drizzle db and its faked table/column definitions were
duplicated (and drifting) between apps/worker and apps/api test setup.
Move it into a shared createDbMock() factory in test-fixtures as the
superset of both, so both apps' vi.mock("@playfulprogramming/db", ...)
calls source the same mock instead of hand-rolling their own.
