+++ apps/worker/test-utils/setup.ts
@@ -1,6 +1,6 @@
 import "./server.ts";
 import { vi, afterEach, beforeEach } from "vitest";
-import "@playfulprogramming/test-fixtures";
+import { createDbMock } from "@playfulprogramming/test-fixtures";
 import { s3 } from "@playfulprogramming/s3";
 import { Readable } from "node:stream";
 
@@ -40,92 +40,7 @@ vi.mock("@playfulprogramming/s3", () => {
 	};
 });
 
-vi.mock("@playfulprogramming/db", () => {
-	const insertMap = new Map<unknown, unknown>();
-	const insertMockResponse = () => {
-		const onConflictDoUpdate = vi.fn();
-		return { values: vi.fn(() => ({ onConflictDoUpdate })) };
-	};
-
-	const deleteMap = new Map<unknown, unknown>();
-	const deleteMockResponse = () => {
-		const returning = vi.fn();
-		return { where: vi.fn(() => ({ returning })) };
-	};
-
-	const db = {
-		insert: vi.fn((arg) => {
-			return (
-				insertMap.get(arg) ?? insertMap.set(arg, insertMockResponse()).get(arg)
-			);
-		}),
-		delete: vi.fn((arg) => {
-			return (
-				deleteMap.get(arg) ?? deleteMap.set(arg, deleteMockResponse()).get(arg)
-			);
-		}),
-		select: vi.fn(),
-		transaction: vi.fn((cb: (tx: unknown) => unknown) => cb(db)),
-	};
-
-	return {
-		profiles: {
-			slug: {},
-		},
-		profileAchievements: {
-			profileSlug: {},
-			achievementId: {},
-		},
-		authorRoles: {
-			profileSlug: {},
-			role: {},
-		},
-		posts: {
-			slug: {},
-		},
-		collections: {
-			slug: {},
-		},
-		collectionData: {
-			slug: {},
-			locale: {},
-		},
-		collectionAuthors: {
-			collectionSlug: {},
-			authorSlug: {},
-		},
-		collectionTags: {
-			collectionSlug: {},
-			tag: {},
-		},
-		postData: {
-			slug: {},
-			locale: {},
-			version: {},
-		},
-		postAuthors: {
-			postSlug: {},
-			authorSlug: {},
-		},
-		postTags: {
-			postSlug: {},
-			tag: {},
-		},
-		postAttachments: {
-			postSlug: {},
-			attachmentName: {},
-		},
-		collectionChapters: {
-			postSlug: {},
-			locale: {},
-		},
-		urlMetadata: {},
-		urlMetadataPost: {},
-		urlMetadataGist: {},
-		urlMetadataGistFile: {},
-		db,
-	};
-});
+vi.mock("@playfulprogramming/db", () => createDbMock());
 
 vi.mock("@playfulprogramming/github-api", async (importOriginal) => {
 	const actual = await importOriginal();
