+++ packages/test-fixtures/src/index.ts
@@ -1,1 +1,3 @@
 import "./env-stubs.ts";
+
+export { createDbMock } from "./db-mock.ts";
